separate content item into multiple files
