+++ src/pages/Volunteer.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 
-import ContentItem from "./components/ContentItem";
+import ContentItem from "./components/content item/ContentItem";
 import { importAllPhotosFromFolder } from "../helpers/Helpers";
 
 function Volunteer() {
