+++ src/pages/components/Modal.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 
 interface ModalProps {
   image: string;
